Edit the update nimbus, and add empty nimbus feature

+++ src/main/java/dev/pollywag/nimbusdrop/service/NimbusService.java
@@ -2,9 +2,13 @@ package dev.pollywag.nimbusdrop.service;
 
 import dev.pollywag.nimbusdrop.dto.respondeDTO.NimbusResponse;
 import dev.pollywag.nimbusdrop.entity.Nimbus;
+import dev.pollywag.nimbusdrop.entity.NimbusSpace;
 import dev.pollywag.nimbusdrop.entity.User;
 import dev.pollywag.nimbusdrop.exception.NimbusNotFoundException;
+import dev.pollywag.nimbusdrop.exception.ResourceOwnershipException;
+import dev.pollywag.nimbusdrop.repository.DropRepository;
 import dev.pollywag.nimbusdrop.repository.NimbusRepository;
+import dev.pollywag.nimbusdrop.repository.NimbusSpaceRepository;
 import dev.pollywag.nimbusdrop.repository.UserRepository;
 import org.apache.commons.io.FileUtils;
 import org.modelmapper.ModelMapper;
@@ -15,6 +19,7 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.imageio.IIOException;
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -26,12 +31,16 @@ public class NimbusService {
     private final UserRepository userRepository;
     private final ModelMapper modelMapper;
     private final FileStorageService fileStorageService;
+    private final DropRepository dropRepository;
+    private final NimbusSpaceRepository nimbusSpaceRepository;
 
-    public NimbusService(NimbusRepository nimbusRepository, UserRepository userRepository, ModelMapper modelMapper, FileStorageService fileStorageService) {
+    public NimbusService(NimbusSpaceRepository nimbusSpaceRepository,DropRepository dropRepository,NimbusRepository nimbusRepository, UserRepository userRepository, ModelMapper modelMapper, FileStorageService fileStorageService) {
         this.nimbusRepository = nimbusRepository;
         this.userRepository = userRepository;
         this.modelMapper = modelMapper;
         this.fileStorageService = fileStorageService;
+        this.dropRepository = dropRepository;
+        this.nimbusSpaceRepository = nimbusSpaceRepository;
     }
 
     public NimbusResponse createNimbus(String nimbusName, String email) {
@@ -48,18 +57,56 @@ public class NimbusService {
     public void deleteNimbus(Long id, String email){
         String userDisplayName = userRepository.findUserDisplayNameByEmail(email).orElseThrow( () -> new UsernameNotFoundException("Account not found: " + email));
         Nimbus nimbus = nimbusRepository.findById(id).orElseThrow( () -> new NimbusNotFoundException("Nimbus not found: " + id));
+        String nimbusOwnerUsername =nimbus.getNimbusSpace().getUser().getUserDisplayName();
+
+        if(!userDisplayName.equals(nimbusOwnerUsername)){
+            throw new ResourceOwnershipException("You are not allowed to modify this resource");
+        }
+
+        if(!nimbus.getDrops().isEmpty()){
+            throw new IllegalArgumentException("All Drops must be empty for this nimbus before deletion");
+        }
 
         String nimbusName = nimbus.getNimbusName();
         fileStorageService.deleteNimbusDirectory(userDisplayName, nimbusName);
         nimbusRepository.delete(nimbus);
     }
 
-    public NimbusResponse updateNimbusName(Long id, String newNimbusName){
+    public String emptyNimbus(Long id, String email){
+        String userDisplayName = userRepository.findUserDisplayNameByEmail(email).orElseThrow( () -> new UsernameNotFoundException("Account not found: " + email));
         Nimbus nimbus = nimbusRepository.findById(id).orElseThrow( () -> new NimbusNotFoundException("Nimbus not found: " + id));
-        User user = nimbus.getNimbusSpace().getUser();
+        NimbusSpace nimbusSpace = nimbus.getNimbusSpace();
+        String nimbusOwnerUsername =nimbus.getNimbusSpace().getUser().getUserDisplayName();
 
-        String userDisplayName = user.getUserDisplayName();
+        if(!userDisplayName.equals(nimbusOwnerUsername)){
+            throw new ResourceOwnershipException("You are not allowed to modify this resource");
+        }
+
+        if(nimbus.getDrops().isEmpty()){
+            throw new IllegalArgumentException("This nimbus is already empty");
+        }
+
+        String nimbusName = nimbus.getNimbusName();
+
+        fileStorageService.emptyNimbusDirectory(userDisplayName,nimbusName);
+        nimbusSpace.setUsedStorageBytes(0L);
+
+        nimbusSpaceRepository.save(nimbusSpace);
+        int totalOfDropsDeleted = dropRepository.deleteAllByNimbusId(id);
+
+        return totalOfDropsDeleted + " drops were deleted";
+    }
+
+    public NimbusResponse updateNimbusName(Long id, String newNimbusName, String email){
+        String userDisplayName = userRepository.findUserDisplayNameByEmail(email).orElseThrow( () -> new UsernameNotFoundException("Account not found: " + email));
+        Nimbus nimbus = nimbusRepository.findById(id).orElseThrow( () -> new NimbusNotFoundException("Nimbus not found: " + id));
+
+        String nimbusOwnerUsername =nimbus.getNimbusSpace().getUser().getUserDisplayName();
         String oldNimbusName = nimbus.getNimbusName();
+
+        if(!userDisplayName.equals(nimbusOwnerUsername)){
+            throw new ResourceOwnershipException("You are not allowed to modify this resource");
+        }
 
         try {
             fileStorageService.nimbusRename(userDisplayName, newNimbusName, oldNimbusName);
